app.rs clean up
